[base-trigger] renamed story
